Small refactor of model classes
