+++ lib/model/hive/report.dart
